Import plus propre du logo
