+++ source/images/Logo.tsx
@@ -1,0 +1,27 @@
+import * as React from "react";
+
+function SvgLogoNgcGithub(props) {
+  return (
+    <svg
+      width="210mm"
+      height="210mm"
+      viewBox="0 0 210 210"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        fill="none"
+        stroke="var(--color)"
+        strokeWidth={9.80622}
+        d="M8.461 8.225h194.032v194.032H8.461z"
+      />
+      <path
+        d="M10.609 102.554c21.781-35.27 42.728-76.68 65.87-76.68 34.665 0 25.355 116.617 57.514 160.274 7.725 10.487 24.557 12.742 63.566 11.46 0 6.705-136.154 3.843-189.381 3.843z"
+        fill="var(--color)"
+      />
+    </svg>
+  );
+}
+
+export default SvgLogoNgcGithub;
+
