+++ source/sites/publicodes/App.tsx
@@ -18,7 +18,7 @@ import Contribution from './Contribution'
 import Diffuser from './Diffuser'
 import Fin from './Fin'
 import Landing from './Landing'
-import Logo from './Logo'
+import Logo from './LogoComponent'
 import Navigation from './Navigation'
 import Documentation from './pages/Documentation'
 import Personas from './Personas.tsx'
